todoService moded to typeSript - app is typesafe :)

+++ src/services/todoService.ts
@@ -1,26 +1,28 @@
 import axios from "axios";
+import { Todo } from "../store/todos/todoType";
 
-const url = "http://localhost:3001/todos";
+const url: string = "http://localhost:3001/todos";
 
 export const getTodos = async () => {
   const response = await axios.get(url);
   return response.data;
 };
 
-export const addTodo = async todoItem => {
+export const addTodo = async (todoItem: Todo) => {
   const response = await axios.post(url, todoItem);
   console.log(response);
   return response.data;
 };
 
-export const updateTodo = async todoToChange => {
+export const updateTodo = async (todoToChange: Todo) => {
   const urlOfUpdate = url + `/${todoToChange.id}`;
   const response = await axios.put(urlOfUpdate, todoToChange);
   console.log(response);
 };
 
-export const deleteTodo = async id => {
-  const urlOfDelete = url + `/${id}`;
+export const deleteTodo = async (todo: Todo) => {
+  if (todo.id === undefined) throw "Todo has no id :/";
+  const urlOfDelete = url + `/${todo.id}`;
   const response = await axios.delete(urlOfDelete);
   console.log(response);
 };
